feat: split into components
